postprocess: Work around SELinux issue with cross labeling
Unfortunately when composing from RHEL7 a current rawhide tree, the
PCRE version mismatch causes none of the regexps to match...

Work around this by temporarily moving the .bin files out of the way.
This is really a hack though.

The real fix is to change libselinux to check pcre_version().