postprocess: provide --gzip and --reproducible to dracut
This should help to generate the same initrd when the files didn't
change.

Newer versions of gzip (or pigz when available) can generate rsync
friendly files and if present, Dracut already takes advantage of it.

Also use --reproducible, to instruct Dracut to generate CPIO
reproducible files.  It is required a version of GNU CPIO that
has support for it.

Dracut has --reproducible for almost a year, so use it inconditionally.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>