autogen.sh: tweak program checking logic
With the errexit bash option turned on, these conditionals would never
actually be reached since the failure from `which` would cause the
script to exit.

As a result, if autoreconf was not installed, all the user would see
would be the error message from `which`, and not pretty error we have
for them. Similarly, even though gtk-doc should be optional, the script
would fail if gtkdocize wasn't installed.

Also fix minor typo.