core: Print number of relabels that occurred
I'd find this to be very interesting information.  It would guide
things like whether or not we need multiple cached packages per policy
and things like that.

NOTE: I only compile tested this.  It's clear we're going to need real
test coverage that uses different policies and installs packages and
stuff.