Auto upgrading with a script & service
This service file is tested. And it runs the autoupdate.sh file automatically, and check for updates. But the problem is after restarting the VM, autoupdate.sh file was missing. But the service is running. Now I need to find a method to keep autoupdate.sh file permenently even after a reboot.