Makefile-hif.am: link by force and add clean hook
Add a clean hook so that `make clean` on the top dir also cleans the
libhif artifacts.

Add -f to `ln` -- Make marks the target as out-of-date if the symlink
breaks (which will happen if you just `make clean` in the libhif-build
dir), but the `ln` will cause the build to fail since it already exists.